+++ common/src/main/java/com/codinglitch/simpleradio/CompatCore.java
@@ -1,15 +1,9 @@
 package com.codinglitch.simpleradio;
 
 import com.codinglitch.simpleradio.compat.VibrativeCompat;
-import com.codinglitch.simpleradio.core.central.WorldlyPosition;
 import com.codinglitch.simpleradio.platform.Services;
 import com.codinglitch.simpleradio.radio.RadioChannel;
 import com.codinglitch.simpleradio.radio.RadioSource;
-import net.minecraft.core.BlockPos;
-import net.minecraft.world.level.Level;
-import org.joml.Vector3f;
-
-import static com.codinglitch.simpleradio.platform.Services.COMPAT;
 
 public class CompatCore {
     public static boolean VC_INTERACTION = false;
@@ -17,12 +11,10 @@ public class CompatCore {
     public static boolean VALKYRIEN_SKIES = false;
 
     public static void spoutCompatibilities() {
-        //TODO: add a reload method from lexiconfig so we can actually use the fields above
-
         //---- Voice Chat Interaction ----\\
         if (Services.PLATFORM.isModLoaded("vcinteraction")) {
             CommonSimpleRadio.info("Voice Chat Interaction is present!");
-            if (CommonSimpleRadio.SERVER_CONFIG.compatibilities.voice_chat_interaction.enabled) {
+            if (SimpleRadioLibrary.SERVER_CONFIG.compatibilities.voice_chat_interaction.enabled) {
                 VC_INTERACTION = true;
                 CommonSimpleRadio.info("..and compat is enabled!");
             } else {
@@ -36,7 +28,7 @@ public class CompatCore {
             if (Services.PLATFORM.isModLoaded("vcinteraction")) {
                 CommonSimpleRadio.info("..but so is Voice Chat Interaction?!");
             } else {
-                if (CommonSimpleRadio.SERVER_CONFIG.compatibilities.vibrative_voice.enabled) {
+                if (SimpleRadioLibrary.SERVER_CONFIG.compatibilities.vibrative_voice.enabled) {
                     VIBRATIVE_VOICE = true;
                     CommonSimpleRadio.info("..and compat is enabled!");
                 } else {
@@ -48,7 +40,7 @@ public class CompatCore {
         //---- Valkyrien Skies ----\\
         if (Services.PLATFORM.isModLoaded("valkyrienskies")) {
             CommonSimpleRadio.info("Valkyrien Skies is present!");
-            if (CommonSimpleRadio.SERVER_CONFIG.compatibilities.valkyrien_skies.enabled) {
+            if (SimpleRadioLibrary.SERVER_CONFIG.compatibilities.valkyrien_skies.enabled) {
                 VALKYRIEN_SKIES = true;
                 CommonSimpleRadio.info("..and compat is enabled!");
             } else {
@@ -57,9 +49,14 @@ public class CompatCore {
         }
     }
 
+    public static void reloadCompatibilities() {
+        CommonSimpleRadio.info("Reloading compatibilities!");
+        spoutCompatibilities();
+    }
+
     public static void onData(RadioChannel channel, RadioSource source, short[] decoded) {
         // ---- Vibrative Voice ---- \\
-        if (Services.PLATFORM.isModLoaded("vibrativevoice") && CommonSimpleRadio.SERVER_CONFIG.compatibilities.voice_chat_interaction.enabled) {
+        if (CompatCore.VIBRATIVE_VOICE) {
             VibrativeCompat.onData(channel, source, decoded);
         }
     }