+++ common/src/main/java/com/codinglitch/simpleradio/core/registry/blocks/SpeakerBlockEntity.java
@@ -4,6 +4,7 @@ import com.codinglitch.simpleradio.CompatCore;
 import com.codinglitch.simpleradio.core.central.Frequency;
 import com.codinglitch.simpleradio.core.central.FrequencyBlockEntity;
 import com.codinglitch.simpleradio.core.central.Receiving;
+import com.codinglitch.simpleradio.core.central.WorldlyPosition;
 import com.codinglitch.simpleradio.core.registry.SimpleRadioBlockEntities;
 import com.codinglitch.simpleradio.core.registry.SimpleRadioSounds;
 import com.codinglitch.simpleradio.platform.Services;
@@ -32,10 +33,9 @@ public class SpeakerBlockEntity extends FrequencyBlockEntity implements Receivin
 
     @Override
     public void setRemoved() {
-        if (level != null && !level.isClientSide) {
-            Vector3f locationVec = Services.COMPAT.modifyPosition(level, this.worldPosition);
+        if (level != null && !level.isClientSide && this.channel != null) {
             level.playSound(
-                    null, locationVec.x, locationVec.y, locationVec.z,
+                    null, channel.location.x, channel.location.y, channel.location.z,
                     SimpleRadioSounds.RADIO_CLOSE,
                     SoundSource.PLAYERS,
                     1f, 1f
@@ -67,7 +67,6 @@ public class SpeakerBlockEntity extends FrequencyBlockEntity implements Receivin
 
     public static void tick(Level level, BlockPos pos, BlockState blockState, SpeakerBlockEntity blockEntity) {
         if (!level.isClientSide) {
-            if (blockEntity.channel != null) { blockEntity.channel.location = Services.COMPAT.modifyPosition(pos, level); }
             if (blockEntity.frequency != null && !blockEntity.isListening) {
                 blockEntity.listen();
             }
@@ -76,11 +75,10 @@ public class SpeakerBlockEntity extends FrequencyBlockEntity implements Receivin
 
     public void listen() {
         channel = startReceiving(frequency.frequency, frequency.modulation, listenerID);
-        channel.location = Services.COMPAT.modifyPosition(this.worldPosition, this.level);
+        channel.location = Services.COMPAT.modifyPosition(Services.COMPAT.modifyPosition(WorldlyPosition.of(worldPosition, level, worldPosition)));
 
-        Vector3f locationVec = Services.COMPAT.modifyPosition(level, this.worldPosition);
         level.playSound(
-                null, locationVec.x, locationVec.y, locationVec.z,
+                null, channel.location.x, channel.location.y, channel.location.z,
                 SimpleRadioSounds.RADIO_OPEN,
                 SoundSource.PLAYERS,
                 1f, 1f